Renamed Mixins.
Try to make mixins use inheritance a little better. This should allow for additional mixins to be easily configured.

When a setting is loaded via the _shareMixin, change the parameter name to prevent the loaded settings from overriding the user's existing settings. The user can still make changes and share the map, but when they reload, the settings will be restored to their previous state.